Break out parse_roles (and its validation) into a separate module
