+++ tests/jobserver/authorization/test_fields.py
@@ -1,42 +1,10 @@
 import pytest
 from django.db import connection
 
-from jobserver.authorization import CoreDeveloper, OutputChecker, ProjectCollaborator
-from jobserver.authorization.fields import RolesField, _ensure_role_paths, parse_roles
+from jobserver.authorization import CoreDeveloper
+from jobserver.authorization.fields import RolesField
 
 from ...factories import OrgFactory, OrgMembershipFactory, UserFactory
-
-
-def test_ensure_role_paths_success():
-    paths = [
-        "jobserver.authorization.roles.OutputChecker",
-        "jobserver.authorization.roles.ProjectCollaborator",
-    ]
-
-    assert _ensure_role_paths(paths) is None
-
-
-def test_ensure_role_paths_with_invalid_paths():
-    paths = [
-        "jobserver.authorization.roles.OutputChecker",
-        "test.dummy.SomeRole",
-    ]
-
-    msg = "Some Role paths did not start with jobserver.authorization.roles:\n"
-    msg += " - test.dummy.SomeRole"
-    with pytest.raises(ValueError, match=msg):
-        _ensure_role_paths(paths)
-
-
-def test_parse_roles_success():
-    paths = [
-        "jobserver.authorization.roles.OutputChecker",
-        "jobserver.authorization.roles.ProjectCollaborator",
-    ]
-
-    roles = parse_roles(paths)
-
-    assert roles == [OutputChecker, ProjectCollaborator]
 
 
 @pytest.mark.django_db
